db updates, use only messages sent by self in db

+++ pgbot/db.py
@@ -20,7 +20,8 @@ db_channel = None
 # Store "name: pickled data" pairs as cache. Do not store unpickled data
 db_obj_cache = {}
 
-write_lock = False
+# We need a lock because we do not want simultaneous access to DB
+lock = asyncio.Lock()
 
 
 async def init(db_chan: discord.TextChannel):
@@ -32,11 +33,16 @@ async def init(db_chan: discord.TextChannel):
     db_channel = db_chan
 
     async for msg in db_channel.history():
-        if msg.attachments:
+        if msg.author.id == common.bot.user.id and msg.attachments:
             db_obj_cache[msg.content] = await msg.attachments[0].read()
 
 
 class DiscordDB:
+    """
+    DiscordDB is a class to interface with a DB like solution, that stores data
+    via discord messages
+    """
+
     def __init__(self, name: str):
         """
         Initialise Discord DB Object
@@ -47,39 +53,36 @@ class DiscordDB:
         """
         Get object of discord DB
         """
-        if common.TEST_MODE:
-            # Caches are not reliable on test mode, when more than one testbot runs
-            async for msg in db_channel.history():
-                if msg.attachments and msg.content == self.name:
-                    db_obj_cache[msg.content] = await msg.attachments[0].read()
-                    break
+        async with lock:
+            if common.TEST_MODE:
+                # Caches are not reliable on test mode, when more than one
+                # testbot runs
+                async for msg in db_channel.history():
+                    if msg.author.id != common.bot.user.id:
+                        continue
 
-        try:
-            return pickle.loads(db_obj_cache[self.name])
-        except KeyError:
-            return failobj
+                    if msg.attachments and msg.content == self.name:
+                        db_obj_cache[self.name] = await msg.attachments[0].read()
+                        break
+
+            try:
+                return pickle.loads(db_obj_cache[self.name])
+            except KeyError:
+                return failobj
 
     async def write(self, obj):
         """
         Store object in DB
         """
-        global write_lock
 
-        while write_lock:
-            await asyncio.sleep(0.002)
+        async with lock:
+            async for msg in db_channel.history():
+                if msg.author.id != common.bot.user.id:
+                    continue
 
-        write_lock = True
-        try:
-            with io.BytesIO() as fobj:
-                pickle.dump(obj, fobj)
-                fobj.seek(0)
-                db_obj_cache[self.name] = fobj.read()
+                if msg.content == self.name:
+                    await msg.delete()
 
-                async for msg in db_channel.history():
-                    if msg.content == self.name:
-                        await msg.delete()
-
-                fobj.seek(0)
+            db_obj_cache[self.name] = pickle.dumps(obj)
+            with io.BytesIO(db_obj_cache[self.name]) as fobj:
                 await db_channel.send(self.name, file=discord.File(fobj))
-        finally:
-            write_lock = False
